Add files via upload
This file explores how 3 interconnected games interact over repeated rounds when each game has a maximum sustainable cost, and the costs accrue round by round. The cost for a game is determined by the game outcome and the game outcome of the other games. This shows that if one game is in a Nash Equilibrium all connected games will be in a Nash Equilibrium.